Move path to better location
