+++ lib/hanami/routing/http_router.rb
@@ -6,7 +6,7 @@ require 'hanami/routing/parsers'
 require 'hanami/routing/force_ssl'
 require 'hanami/routing/error'
 require 'hanami/utils/path_prefix'
-require 'hanami/http_router_monkey_patch'
+require 'hanami/routing/http_router_monkey_patch'
 
 module Hanami
   module Routing
